+++ .github/workflows/deploy-pr.yml
@@ -60,8 +60,9 @@ jobs:
 
       - name: Deploy PR
         env:
-          MDBLIST_API_KEY: ${{ secrets.MDBLIST_API_KEY }}
+          API_KEY_WEB: ${{ secrets.API_KEY_WEB }}
           AWS_HOSTED_ZONE_ID_DEV: ${{ secrets.AWS_HOSTED_ZONE_ID_DEV }}
+          MDBLIST_API_KEY: ${{ secrets.MDBLIST_API_KEY }}
           SECRET_KEY: ${{ secrets.SECRET_KEY }}
           TMDB_API_ACCESS_TOKEN: ${{ secrets.TMDB_API_ACCESS_TOKEN }}
           TMDB_API_KEY: ${{ secrets.TMDB_API_KEY }}
