+++ apps/api/src/lambdas/dominantColor.ts
