+++ apps/api/src/lambdas/validateWatchedStatusCron.ts
