+++ apps/api/src/lib/db/webhooks/index.ts
@@ -6,16 +6,10 @@ import {
   QueryCommand,
 } from '@aws-sdk/client-dynamodb';
 import { marshall, unmarshall } from '@aws-sdk/util-dynamodb';
+import { type WebhookToken } from '@tvseri.es/types';
 import { Resource } from 'sst';
 
 import client from '../client';
-
-type WebhookToken = Readonly<{
-  token: string;
-  userId: string;
-  type: string;
-  createdAt: string;
-}>;
 
 const generateToken = () => {
   return randomBytes(12).toString('hex');
