+++ apps/web/src/components/Watched/AddTvSeriesToStore.tsx
@@ -2,9 +2,7 @@
 
 import { useEffect, useRef } from 'react';
 
-import { type TvSeries } from '@tvseri.es/types';
-
-import { type WatchedItem } from '@/lib/db/watched';
+import { type TvSeries, type WatchedItem } from '@tvseri.es/types';
 
 import { useWatchedStore } from './WatchedStoreProvider';
