+++ apps/web/src/components/Watched/store.ts
@@ -1,8 +1,8 @@
 import { type TvSeries } from '@tvseri.es/types';
+import type { WatchedItem } from '@tvseri.es/types';
 import { createStore, type StateCreator } from 'zustand';
 import { devtools } from 'zustand/middleware';
 
-import { type WatchedItem } from '@/lib/db/watched';
 import calculateProgress from '@/utils/calculateProgress';
 
 type WatchedItemForStore = Pick<
