+++ packages/types/src/cache.ts
@@ -1,0 +1,9 @@
+export type CacheOptions = Readonly<{
+  ttl?: number | null; // null means no expiration
+}>;
+
+export type CacheItem = Readonly<{
+  value: string;
+  expiresAt?: number;
+  createdAt: string;
+}>;
