+++ packages/types/src/index.ts
@@ -1,12 +1,19 @@
 export * from './account';
+export * from './cache';
 export * from './country-language';
 export * from './genre';
 export * from './keyword';
+export * from './list';
 export * from './movie';
+export * from './pagination-sorting';
 export * from './person';
+export * from './preferred-images';
 export * from './rating';
+export * from './scrobble';
 export * from './session';
 export * from './tmdb';
 export * from './tv-series';
 export * from './user';
 export * from './watch-provider';
+export * from './watched';
+export * from './webhook';
