+++ packages/types/src/preferred-images.ts
@@ -1,0 +1,5 @@
+export type PreferredImages = Readonly<{
+  backdropImagePath: string;
+  backdropColor: string;
+  titleTreatmentImagePath?: string;
+}>;
