+++ packages/types/src/webhook.ts
@@ -1,0 +1,6 @@
+export type WebhookToken = Readonly<{
+  token: string;
+  userId: string;
+  type: string;
+  createdAt: string;
+}>;
