+++ sst.config.ts
@@ -23,6 +23,7 @@ export default $config({
     await import('./infra/dynamo');
     await import('./infra/scrobbleQueue');
     await import('./infra/watchedStatus');
+    await import('./infra/api');
     await import('./infra/web');
   },
 });
